// src/types/schema.ts
export type ID = string;
export type SchemaKey = "Dandiset";
export type SchemaVersion = string;
/**
 * A title associated with the Dandiset.
 */
export type DandisetTitle = string;
/**
 * A description of the Dandiset
 */
export type Description = string;
/**
 * People or Organizations that have contributed to this Dandiset.
 */
export type DandisetContributors = [Person | Organization, ...(Person | Organization)[]];
// ...
export type SubjectMatterOfTheDataset = (Disorder | Anatomy | GenericType)[];
/**
 * Objectives or specific questions of the study.
 */
export type StudyTarget = string[];
/**
 * Licenses associated with the item. DANDI only supports a subset of Creative Commons Licenses (creativecommons.org) applicable to datasets.
 */
export type License = [LicenseType, ...LicenseType[]];
/**
 * An enumeration of supported licenses
 */
export type LicenseType = "spdx:CC0-1.0" | "spdx:CC-BY-4.0";
/**
 * A list of persistent URLs describing the protocol (e.g. protocols.io, or other DOIs).
 */
export type Protocol = string[];
/**
 * Uniform resource identifier
 */
export type ID8 = string;
export type SchemaKey8 = "EthicsApproval";
/**
 * Approved Protocol identifier, often a number or alphanumeric string.
 */
export type ApprovedProtocolIdentifier = string;
export type EthicsApprovals = EthicsApproval[];
/**
 * Keywords used to describe this content.
 */
export type Keywords = string[];
export type Acknowledgement = string;
/**
 * Uniform resource identifier
 */
export type ID9 = string;
export type SchemaKey9 = "AccessRequirements";
/**
 * The access status of the item.
 */
export type AccessStatus = "dandi:OpenAccess";
/**
 * Information about access requirements when embargoed or restricted
 */
export type Description1 = string;
/**
 * Date on which embargo ends.
 */
export type EmbargoEndDate = string;
export type AccessInformation = AccessRequirements[];
/**
 * permalink to the item
 */
export type URL3 = string;
/**
 * location of the item
 */
export type Repository = string;
// ...
export type RelatedResource = Resource[];
// ...
export type Version = string;
// ...
export type AssociatedProjects = Project[];
/**
 * A Dandiset identifier that can be resolved by identifiers.org.
 */
export type DandisetIdentifier = string;
export type DandisetCreationDateAndTime = string;
export type LastModificationDateAndTime = string;
export type Citation = string;
/**
 * Uniform resource identifier
 */
export type ID15 = string;
export type SchemaKey15 = "AssetsSummary";
export type NumberOfBytes = number;
export type NumberOfFiles = number;
export type NumberOfSubjects = number;
export type NumberOfSamples = number;
export type NumberOfCells = number;
/**
 * Uniform resource identifier
 */
export type ID16 = string;
export type SchemaKey16 = "StandardsType";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier7 = string;
/**
 * The name of the item.
 */
export type Name8 = string;
export type DataStandard = StandardsType[];
/**
 * Uniform resource identifier
 */
export type ID17 = string;
export type SchemaKey17 = "ApproachType";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier8 = string;
/**
 * The name of the item.
 */
export type Name9 = string;
export type Approach = ApproachType[];
/**
 * Uniform resource identifier
 */
export type ID18 = string;
export type SchemaKey18 = "MeasurementTechniqueType";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier9 = string;
/**
 * The name of the item.
 */
export type Name10 = string;
export type MeasurementTechnique = MeasurementTechniqueType[];
export type VariableMeasured = string[];
/**
 * Uniform resource identifier
 */
export type ID19 = string;
export type SchemaKey19 = "SpeciesType";
/**
 * The identifier can be any url or a compact URI, preferably supported by identifiers.org.
 */
export type Identifier10 = string;
/**
 * The name of the item.
 */
export type Name11 = string;
export type Species = SpeciesType[];
export type ManifestLocation = [string, ...string[]];
export type Version1 = string;

/**
 * A body of structured information describing a DANDI dataset.
 */
export interface Dandiset {
  id: ID;
  schemaKey: SchemaKey;
  schemaVersion?: SchemaVersion;
  name: DandisetTitle;
  description: Description;
  contributor: DandisetContributors;
  about?: SubjectMatterOfTheDataset;
  studyTarget?: StudyTarget;
  license: License;
  protocol?: Protocol;
  ethicsApproval?: EthicsApprovals;
  keywords?: Keywords;
  acknowledgement?: Acknowledgement;
  access?: AccessInformation;
  url?: URL3;
  repository?: Repository;
  relatedResource?: RelatedResource;
  wasGeneratedBy?: AssociatedProjects;
  identifier: DandisetIdentifier;
  dateCreated?: DandisetCreationDateAndTime;
  dateModified?: LastModificationDateAndTime;
  citation: Citation;
  assetsSummary: AssetsSummary;
  manifestLocation: ManifestLocation;
  version: Version1;
  [k: string]: unknown;
}
export interface Person {
  id?: ID1;
  schemaKey: SchemaKey1;
  identifier?: AnORCIDIdentifier;
  name: Name;
  email?: Email;
  url?: URL;
  roleName?: Role;
  includeInCitation?: IncludeContributorInCitation;
  awardNumber?: IdentifierForAnAward;
  affiliation?: Affiliation;
  [k: string]: unknown;
}
export interface Affiliation1 {
  id?: ID2;
  schemaKey: SchemaKey2;
  identifier?: ARorOrgIdentifier;
  name: Name1;
  [k: string]: unknown;
}
export interface Organization {
  id?: ID3;
  schemaKey: SchemaKey3;
  identifier?: ARorOrgIdentifier1;
  name?: Name2;
  email?: Email1;
  url?: URL1;
  roleName?: Role1;
  includeInCitation?: IncludeContributorInCitation1;
  awardNumber?: IdentifierForAnAward1;
  contactPoint?: OrganizationContactInformation;
  [k: string]: unknown;
}
export interface ContactPoint {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}
/**
 * Biolink, SNOMED, or other identifier for disorder studied
 */
export interface Disorder {
  id?: ID5;
  schemaKey: SchemaKey5;
  identifier?: Identifier;
  name?: Name3;
  dxdate?: DatesOfDiagnosis;
  [k: string]: unknown;
}
/**
 * UBERON or other identifier for anatomical part studied
 */
export interface Anatomy {
  id?: ID6;
  schemaKey: SchemaKey6;
  identifier?: Identifier1;
  name?: Name4;
  [k: string]: unknown;
}
/**
 * An object to capture any type for about
 */
export interface GenericType {
  id?: ID7;
  schemaKey: SchemaKey7;
  identifier?: Identifier2;
  name?: Name5;
  [k: string]: unknown;
}
/**
 * Information about ethics committee approval for project
 */
export interface EthicsApproval {
  id?: ID8;
  schemaKey: SchemaKey8;
  identifier: ApprovedProtocolIdentifier;
  contactPoint?: ContactPoint1;
  [k: string]: unknown;
}
/**
 * Information about the ethics approval committee.
 */
export interface ContactPoint1 {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}
/**
 * Information about access options for the dataset
 */
export interface AccessRequirements {
  id?: ID9;
  schemaKey: SchemaKey9;
  status: AccessStatus;
  contactPoint?: ContactPoint2;
  description?: Description1;
  embargoedUntil?: EmbargoEndDate;
  [k: string]: unknown;
}
/**
 * Who or where to look for information about access.
 */
export interface ContactPoint2 {
  id?: ID4;
  schemaKey: SchemaKey4;
  email?: Email2;
  url?: URL2;
  [k: string]: unknown;
}
export interface Resource {
  id?: ID10;
  schemaKey: SchemaKey10;
  identifier?: Identifier3;
  name?: ATitleOfTheResource;
  url?: URLOfTheResource;
  repository?: NameOfTheRepository;
  relation: ResourceRelation;
  [k: string]: unknown;
}
/**
 * Information about the Project activity
 */
export interface Project {
  id?: ID11;
  schemaKey: SchemaKey11;
  identifier?: Identifier4;
  name: NameOfProject;
  description?: Description2;
  startDate?: StartDate;
  endDate?: EndDate;
  wasAssociatedWith?: WasAssociatedWith;
  used?: Used;
  [k: string]: unknown;
}
// ...
export interface AssetsSummary {
  id?: ID15;
  schemaKey: SchemaKey15;
  numberOfBytes: NumberOfBytes;
  numberOfFiles: NumberOfFiles;
  numberOfSubjects?: NumberOfSubjects;
  numberOfSamples?: NumberOfSamples;
  numberOfCells?: NumberOfCells;
  dataStandard?: DataStandard;
  approach?: Approach;
  measurementTechnique?: MeasurementTechnique;
  variableMeasured?: VariableMeasured;
  species?: Species;
  [k: string]: unknown;
}